requirements.txt
convertbng==0.6.10
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
numpy==1.16.2
pandas==0.24.2
psycopg2==2.8.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
